sql-proxy-server: cache kubernetes service lookups
This PR introduces a cache for vtgate service addresses.

Currently we make a kubernetes lookup each time a new connection is made
to `sql-proxy-server`. This means we introduce a latency for each
connection. To avoid the latency, we should cache the remote address for
a particular period (which is a minute right now).