proxy: ensure Client fields are 64-bit aligned
fixes https://github.com/planetscale/project-big-bang/issues/135

It looks like there is a bug in Go's runtime that happens on 32-bit
systems:
https://golang.org/pkg/sync/atomic/#pkg-note-BUG It happens if we use 64-bit
operations on 64-bit field. On 32-bit architectures, we need
to make sure to arrange for 64-bit alignment for atomic access. There is
proposal to improve the situation:
https://github.com/golang/go/issues/36606

This PR moves the `connectionsCounter` field to the top of the struct
and hence makes sure it's guaranteed to be 64-bit aligned.